error handler updated

+++ src/controllers/users.controller.js
@@ -1,6 +1,7 @@
 import { prisma } from "../../prisma/prisma.js";
 import bcrypt from "bcrypt";
 import { asyncHandler } from "../utils/asyncHandler.js";
+import { AppError } from "../utils/AppError.js";
 
 // Get all users
 export const getUsers = asyncHandler(async (req, res) => {
@@ -26,7 +27,7 @@ export const getUsers = asyncHandler(async (req, res) => {
 
 // Get single user
 export const getUser = asyncHandler(async (req, res) => {
-  const id = parseInt(req.params.id);
+  const id = req.params.id;
 
   const user = await prisma.user.findUnique({
     where: { id },
@@ -40,10 +41,7 @@ export const getUser = asyncHandler(async (req, res) => {
   });
 
   if (!user) {
-    return res.status(404).json({
-      success: false,
-      message: "User not found",
-    });
+    throw new AppError("User not found", 404);
   }
 
   res.json({
@@ -54,8 +52,16 @@ export const getUser = asyncHandler(async (req, res) => {
 
 // Update user
 export const updateUser = asyncHandler(async (req, res) => {
-  const id = parseInt(req.params.id);
+  const id = req.params.id;
   const { full_name, email, password, role } = req.body;
+
+  const user = await prisma.user.findUnique({
+    where: { id },
+  });
+
+  if (!user) {
+    throw new AppError("User not found", 404);
+  }
 
   const data = {};
 
@@ -89,6 +95,14 @@ export const updateUser = asyncHandler(async (req, res) => {
 // Delete user
 export const deleteUser = asyncHandler(async (req, res) => {
   const id = req.params.id;
+
+  const user = await prisma.user.findUnique({
+    where: { id },
+  });
+
+  if (!user) {
+    throw new AppError("User not found", 404);
+  }
 
   await prisma.user.delete({
     where: { id },
